face_service/requirements.txt
fastapi
uvicorn[standard]
python-multipart
psycopg2-binary
numpy
pillow
face_recognition